separate Action into actions and cleanup namespace
